+++ src/components/layout/mode-switch.tsx
@@ -7,22 +7,20 @@ import { Icons } from '@/components/icons';
 import { Switch } from '@/components/ui/switch';
 
 export function ModeSwitch() {
-  const { theme, setTheme } = useTheme();
+  const { resolvedTheme, setTheme } = useTheme();
   const [mounted, setMounted] = useState(false);
 
   useEffect(() => {
     setMounted(true);
   }, []);
 
-  if (!mounted) {
-    return null;
-  }
+  if (!mounted) return null;
 
   return (
     <div className="flex items-center space-x-2">
       <Icons.sun className="h-[1.2rem] w-[1.2rem]" />
       <Switch
-        checked={theme === 'dark'}
+        checked={resolvedTheme === 'dark'}
         onCheckedChange={(checked) => {
           setTheme(checked ? 'dark' : 'light');
         }}
